Backend/biat_testmanager/pyproject.toml
# pyproject.toml
[project]
name = "biat-testmanager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=4.2,<5.0",
    "djangorestframework>=3.14.0",
    "channels>=4.1.0",
    "channels-redis>=4.2.0",
    "daphne>=4.1.2",
    "python-decouple>=3.8",
    "psycopg2-binary>=2.9.9",
    "celery[redis]>=5.5.3",
    "croniter>=6.0.0",
    "pgvector>=0.4.2",
    "django-cors-headers>=4.9.0",
    "djangorestframework-simplejwt>=5.5.1",
    "django-encrypted-model-fields>=0.6.5",
    "pillow>=12.1.1",
    "playwright>=1.55.0",
    "selenium>=4.25.0",
    "torch>=2.8.0",
    "sentence-transformers>=5.1.0",
    "mlflow>=3.1.0",
    "psutil>=7.1.0",
    "pynvml>=12.0.0",
    "openpyxl>=3.1.5",
    "pypdf>=5.4.0",
    "python-docx>=1.1.2",
    "langgraph>=0.4.8",
]
